refactor: move out theme dropdown component
