+++ components/Header.tsx
@@ -2,9 +2,9 @@ import Link from "next/link";
 import React, { ReactNode } from "react";
 import { useSession, signIn, signOut } from "next-auth/react";
 import styles from "./Header.module.css";
-import { LogoIcon } from "@/components/icons";
-import { ThemeDropdown } from "@/components/layout/Dock/DockNavigation";
+import { ThemeDropdown } from "@/components/ThemeDropdown";
 import { CSSProperties } from 'react';
+import { GitHubIcon, MailIcon } from "@/components/icons";
 
 // `Header` navigation top level component.
 //
@@ -15,7 +15,7 @@ import { CSSProperties } from 'react';
 // rendering, and avoids any flash incorrect content on initial page load.
 export default function Header() {
   return (
-    <header className="z-50 py-4 ">
+    <header className="z-50 py-0.5 shadow-md dark:shadow-on-secondary/40">
       <noscript>
         <style>{`.nojs-show { opacity: 1; top: 0; }`}</style>
       </noscript>
@@ -26,10 +26,8 @@ export default function Header() {
         </div>
       </div>
 
-      <nav className="py-2 font-bold">
-        <div className="opacity-95 dark:prose-a:text-green-300/80">
-          <Navigation />
-        </div>
+      <nav className="py-2 font-bold  prose-a:prose-sm font-sans">
+        <Navigation />
       </nav>
     </header>
   );
@@ -116,13 +114,13 @@ const LogoImage = ({ src, alt }: LogoImageProps) => {
 
 function Navigation() {
   return (
-    <div className="mx-auto px-6 font-sans">
+    <div className="mx-auto px-6">
       <div className="flex items-center gap-x-2 dark:prose-a:font-extrabold dark:prose-a:text-green-300 gap-y-4 w-full justify-between  ">
         <div className="flex gap-8 w-full">
           <Link href="/" data-id="logo" className="mr-auto">
             <LogoImage src={'/favicon-32x32.png'} alt="lloydlobo.com logo" />
           </Link>
-          <div className="flex items-center gap-x-3 flex-1">
+          <div className="flex items-center tracking-wide gap-x-3 flex-1">
             <Link href="/about">About</Link>
             <Link href="/crafts">Crafts</Link>
             <Link href="/projects">Projects</Link>
@@ -168,36 +166,3 @@ function Navigation() {
   );
 }
 
-const GitHubIcon = () => {
-  return (
-    <svg viewBox="0 0 24 24" className="h-6 w-6">
-      <path
-        fill="none"
-        stroke="currentColor"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        strokeWidth="2"
-        d="M9 19c-4.3 1.4-4.3-2.5-6-3m12 5v-3.5c0-1 .1-1.4-.5-2 2.8-.3 5.5-1.4 5.5-6a4.6 4.6 0 0 0-1.3-3.2 4.2 4.2 0 0 0-.1-3.2s-1.1-.3-3.5 1.3a12.3 12.3 0 0 0-6.2 0C6.5 2.8 5.4 3.1 5.4 3.1a4.2 4.2 0 0 0-.1 3.2A4.6 4.6 0 0 0 4 9.5c0 4.6 2.7 5.7 5.5 6-.6.6-.6 1.2-.5 2V21"
-      ></path>
-    </svg>
-  );
-};
-
-const MailIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fill="none"
-      viewBox="0 0 24 24"
-      strokeWidth={1.5}
-      stroke="currentColor"
-      className="h-6 w-6"
-    >
-      <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M21.75 6.75v10.5a2.25 2.25 0 01-2.25 2.25h-15a2.25 2.25 0 01-2.25-2.25V6.75m19.5 0A2.25 2.25 0 0019.5 4.5h-15a2.25 2.25 0 00-2.25 2.25m19.5 0v.243a2.25 2.25 0 01-1.07 1.916l-7.5 4.615a2.25 2.25 0 01-2.36 0L3.32 8.91a2.25 2.25 0 01-1.07-1.916V6.75"
-      />
-    </svg>
-  );
-};
